Moved duplicated admin analytics files into shared module
no ref

- site and post analytics each had an identical empty-stat-view, differing only by a -mt-10 offset; the shared copy takes a className prop and the site call site passes the offset, so both surfaces render exactly as before
- POST_ANALYTICS_INCLUDE now lives with the rest of the shared analytics constants; it had no counterpart or conflict there
- dropped the unused named GrowthSources exports (all call sites import the default); the two growth-sources components themselves stay separate as they intentionally differ per surface
- the site/post KPI_METRICS records and the Stats date-range-select wrapper are deliberately left in place: the records encode surface-specific metrics and colors (visit-duration vs time-on-page), and the wrapper meaningfully binds the shared select to the site analytics context
